requirements.txt
grpcio==1.56.2
grpcio-tools==1.56.2
joblib==1.3.1
numpy==1.25.1
protobuf==4.23.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
scikit-learn==1.3.0
scipy==1.11.1
threadpoolctl==3.2.0
